pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "poisoning_detector"
version = "2.0.0"
requires-python = ">=3.12"
dependencies = [
"beautifulsoup4==4.14.2",
"polars==1.35.1",
"python-dotenv==1.2.1",
"requests==2.32.5",
"web3==7.8.0",
"zstandard==0.25.0",
"psutil==7.1.3",
]

[project.optional-dependencies]
dev = ["pytest>=9.0"]

[tool.hatch.build.targets.wheel]
packages = ["poisoning_detector"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

